all-examples/requirements.txt
pypdf
python-docx
